+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v6.75](https://github.com/replicant4j/replicant/tree/v6.75) (2020-03-04) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.74...v6.75)
 
 * Upgrade the `org.realityforge.braincheck` artifact to version `1.26.0`.